moved the Exceptions and renamed them with capital extracted the appendTokensToCookie method in auth.service.ts
