+++ packages/backend/src/auth/errors/UnauthorisedException.ts
@@ -1,4 +1,4 @@
-import { BaseError } from './baseError'
+import { BaseError } from '../../errors/BaseError'
 export class UnauthorisedException extends BaseError {
   constructor(message: string) {
     super(401, message)
