+++ packages/backend/src/errors/BadRequestException.ts
@@ -1,4 +1,4 @@
-import { BaseError } from './baseError'
+import { BaseError } from './BaseError'
 export class BadRequestException extends BaseError {
   constructor(message: string) {
     super(400, message)
