+++ packages/backend/src/errors/BaseError.ts
