+++ packages/backend/src/middlewares/validator.ts
@@ -1,6 +1,6 @@
 import { Request } from 'express'
 import { AnyZodObject, z, ZodError } from 'zod'
-import { BadRequestException } from '../errors/badRequestException'
+import { BadRequestException } from '../errors/BadRequestException'
 
 export async function zParse<T extends AnyZodObject>(
   schema: T,
